Moved user state handling from main.jsx to Home.jsx
Main.jsx was previously responsible for managing the state and initial loading of the user data. This responsibility has been moved to Home.jsx to make user data more readily available to the components that need it, promoting better data flow and making the application easier to reason about. The related rendering of user details is also reflected in