Uncomment user profile display in Start and pass user object
Uncommented the code for displaying user profile on the 'Start' page which was previously commented out. Passed 'user' object as a prop to the Start component in main.jsx. This is done to display user information right from the start of the application. Also, the 'data' state in Home.jsx is commented as it was not in use.