Temporarily comment out question section
The block of code that displays the current question along with radio buttons for the available options has been commented out. This is for testing purposes and to temporarily bypass the quiz portion while focusing on debugging other functionalities of the Start component. We'll uncomment and restore this part once those issues are resolved.