Update InlinePicker to accept dynamic values
The InlinePicker component is updated to accept dynamic values ('selectionsValue') from its parent component. Previously, the value was hardcoded which limited its reusability and adaptability. Now, it can provide flexibility based on the inputs it receives from the parent components. This change impacts the 'Start' and 'Home' components where the InlinePicker is utilized, therefore the props passed to this component are updated correspondingly in these files.