Add console.log for date difference calculation in Home.jsx
A console.log statement has been added to the getDaysDiff function in Home.jsx component. This change was made to assist in tracking and debugging issues related to the computation of days difference between two dates inside the application. It can help to ensure that the date difference calculations are accurate and are passing the right parameters for inspection.