Refactor calculation of fertility period
The changes made were aimed at increasing the accuracy of the determination of the fertility period in the menstrual cycle. The calculation has now been changed to work with dates, and the "arrayContainsDate" function has been markedly simplified. The update also entails a complete revamp of the conditional statements that used to calculate the